// assets/arc-0031/client/types.ts
export interface Notification {
  type: 'success' | 'warning' | 'error'
  content: string
}

export interface AuthMessage {
  /** The domain name of the Verifier */
  domain: string
  /** Algorand account to authenticate with */
  authAcc: string
  /** Unique random nonce generated by the Verifier */
  nonce: string
  /** Optional, description of the Verifier */
  desc?: string
  /** Optional, metadata */
  meta?: string
}

export interface Session {
  authAcc: string
  accessToken: string
}

export interface AuthStore {
  notification: Notification | null
  message: string
  authMessage: AuthMessage | null
  session: Session | null
}